# kask/docs/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "hKask API",
    "description": "A Minimal Viable Container for Replicants — HTTP API.\n\nhKask is an agent runtime grounded in 12 architectural principles\n(P0–P12) expressed through four composable patterns: Skills Model,\nRegulation Feedback Loop, Agentic AI Mediation, and Agent Creation with\nSovereign Memory. This API exposes all capabilities equally across\nCLI, API, and MCP surfaces under P3 (Equal Surface Exposure).\n\nAll endpoints carry OCAP DelegationToken authentication (P4).\nData access is governed by user sovereignty and affirmative\nconsent (P1, P2).",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.30.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "hKask API server"
    }
  ],
  "paths": {
...
    "/api/chat/stream": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Stream chat inference output as Server-Sent Events.",
        "description": "Accepts the same `ChatRequest` as `/api/chat` but streams the inference\noutput as SSE `data:` events. Each event carries a JSON object with\n`text_delta`, `model`, and optionally `finish_reason`. The final event\nhas `finish_reason: \"stop\"` and includes `usage` stats.\n\nThis is a surface-layer endpoint — it bypasses ChatService's full\npipeline (memory recall, episodic storage) and calls the inference port\ndirectly for streaming. Use `/api/chat` for the full pipeline with\nmemory integration.",
        "operationId": "chat_stream",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "SSE stream of chat chunks",
            "content": {
              "text/event-stream": {}
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/v1/git/archive": {
      "post": {
        "tags": [
          "git"
        ],
        "summary": "Archive a repository template crate",
        "description": "Uses `TemplateCrateLoader::load_template_crate` for template loading (domain\noperation) and `GixCasAdapter::resolve_ref` for SHA resolution.",
        "operationId": "archive",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArchiveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Template crate archived",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArchiveResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid path or missing crate"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/git/resolve/{sha}": {
      "get": {
        "tags": [
          "git"
        ],
        "summary": "Resolve a git reference to a SHA using GixCasAdapter",
        "operationId": "resolve_sha",
        "parameters": [
          {
            "name": "sha",
            "in": "path",
            "description": "Git reference to resolve (branch, tag, or commit prefix)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SHA resolved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveShaResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArchiveEntry": {
        "type": "object",
        "description": "Archive entry response",
        "required": [
          "name",
          "path",
          "content",
          "template_type"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "template_type": {
            "type": "string"
          }
        }
      },
      "ArchiveRequest": {
        "type": "object",
        "description": "Archive repository request",
        "required": [
          "owner",
          "repo",
          "branch",
          "path"
        ],
        "properties": {
          "branch": {
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "repo": {
            "type": "string"
          }
        }
      },
      "ArchiveResponse": {
        "type": "object",
        "description": "Archive response",
        "required": [
          "crate_name",
          "git_sha",
          "templates"
        ],
        "properties": {
          "crate_name": {
            "type": "string"
          },
          "git_sha": {
            "type": "string"
          },
          "templates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArchiveEntry"
            }
          }
        }
      },
...
      "ChatRequest": {
        "type": "object",
        "description": "Chat request sent to the Curator or a specified agent.\n\nThe `model` field allows switching the LLM at request time. When omitted,\nthe server default (a fast classifier model) is used. Use `GET /api/models` to discover\navailable models, and `GET /api/models/search?q=...` for fuzzy matching.",
        "required": [
          "input"
        ],
        "properties": {
          "input": {
            "type": "string",
            "description": "User input message"
          },
          "model": {
            "type": [
              "string",
              "null"
            ],
            "description": "Model override for inference (e.g., \"DI/google/gemma-4-9b-it\"). If unset, uses the server default."
          },
          "template_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional template ID to contextualize the prompt"
          }
        }
      },
...
      "ResolveShaResponse": {
        "type": "object",
        "description": "SHA resolve response",
        "required": [
          "sha"
        ],
        "properties": {
          "sha": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_token": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "DelegationToken",
        "description": "DelegationToken — an OCAP capability token carrying the authenticated WebID and scoped permissions (P4).\n\nObtain via the REPL onboarding flow (`kask secret`) or agent registration (A2A)."
      }
    }
  },
  "security": [
    {
      "bearer_token": []
    }
  ],
...
}